feat(project-creation): convert button to functional component
